update: move factories into its own folder
